Add venue and style improvements
Extracted meeting dates and venue information into a type, but didn't go
so far as to work with Date and Time modules and opted for simple
strings since that's more than enough for what I wanted here.

Added some accessibility improvements with `alt` tags and such.

It took me forevs to find just the right Mickey logo to purchase on Etsy
and drop in as an SVG icon. Worth it.